+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - widgets
   - switch: add `traversing` option (`true` by default)
   - knob: act as a relative encoder when traversing gestures are enabled and `snap` is `false`
+  - keys: `key` variable in `keydown` and `keyup` is now always lowercased when referencing a single character
 - custom module
   - expose shorthand functions `send` and `receives`
   - expose a client object instead of just an id in `app` event handlers to allow retrieving the client's ip address as well